Problems adding hosts to SNO cluster 4.14
