+++ libs/ui-lib/lib/ocm/components/index.ts
@@ -6,3 +6,4 @@ export * from './AddHosts';
 export * from './clusterWizard';
 export * from './HostsClusterDetailTab';
 export * from './featureSupportLevels';
+export * from './newFeatureSupportLevels';
