+++ libs/ui-lib/lib/ocm/components/newFeatureSupportLevels/index.ts
@@ -1,2 +1,3 @@
 export { default as NewFeatureSupportLevelProvider } from './NewFeatureSupportLevelProvider';
 export { default as ClusterNewFeatureSupportLevelsDetailItem } from './ReviewClusterNewFeatureSupportLevels';
+export * from './types';
