+++ libs/ui-lib/lib/ocm/components/newFeatureSupportLevels/types.ts
@@ -1,0 +1,12 @@
+import {
+  ArchitectureSupportLevelMap,
+  NewFeatureSupportLevelMap,
+} from '../../../common/components/newFeatureSupportLevels';
+
+export interface ArchitecturesSupportsLevel {
+  architectures: ArchitectureSupportLevelMap;
+}
+
+export interface FeaturesSupportsLevel {
+  features: NewFeatureSupportLevelMap;
+}
