+++ libs/ui-lib/lib/ocm/services/apis/NewFeatureSupportLevelsAPI.ts
@@ -1,26 +1,24 @@
-import {
-  ArchitectureSupportLevelMap,
-  NewFeatureSupportLevelMap,
-} from '../../../common/components/newFeatureSupportLevels';
+import { CpuArchitecture } from '../../../common';
 import { clientWithoutConverter } from '../../api/axiosClient';
-
-export interface ArchitecturesSupportsLevel {
-  architectures: ArchitectureSupportLevelMap;
-}
-
-export interface FeaturesSupportsLevel {
-  features: NewFeatureSupportLevelMap;
-}
+import {
+  ArchitecturesSupportsLevel,
+  FeaturesSupportsLevel,
+} from '../../components/newFeatureSupportLevels/types';
+import { mapOcmArchToCpuArchitecture } from '../CpuArchitectureService';
 
 const NewFeatureSupportLevelsAPI = {
   makeBaseURI() {
     return `/v2/support-levels`;
   },
 
-  featuresSupportLevel(openshiftVersion: string, cpuArchitecture?: string) {
+  featuresSupportLevel(openshiftVersion: string, cpuArchitecture?: string, isOcm?: boolean) {
+    const cpuArch =
+      isOcm && cpuArchitecture
+        ? mapOcmArchToCpuArchitecture(cpuArchitecture)
+        : cpuArchitecture || CpuArchitecture.x86;
     let queryParams = '?';
     queryParams += openshiftVersion ? `openshift_version=${openshiftVersion}&` : '';
-    queryParams += cpuArchitecture ? `cpu_architecture=${cpuArchitecture}` : '';
+    queryParams += cpuArchitecture ? `cpu_architecture=${cpuArch || CpuArchitecture.x86}` : '';
     return clientWithoutConverter.get<FeaturesSupportsLevel>(
       `${NewFeatureSupportLevelsAPI.makeBaseURI()}/features${queryParams}`,
     );
