+++ libs/ui-lib/lib/ocm/services/apis/index.ts
@@ -6,3 +6,4 @@ export { default as HostsAPI } from './HostsAPI';
 export { default as InfraEnvsAPI } from './InfraEnvsAPI';
 export { default as EventsAPI } from './EventsAPI';
 export { default as FeatureSupportLevelsAPI } from './FeatureSupportLevelsAPI';
+export { default as NewFeatureSupportLevelsAPI } from './NewFeatureSupportLevelsAPI';
